Added ORM to IoTAction.
Added lombok to pom.xml.
IoTActionDAO is the data access object for all IoTActions in the database.

Used lombok on IoTAction to auto-generate setters and getters.

Updated IoTActionList to accept a list in the constructor.

Added shutdown hook in Main.

PathConverter is needed for ScriptIoTAction to convert the Path object (an actual file) to a string and vice-versa.

Updated persistence.xml to use ZIoT as the name and set hibernate to update the database rather than recreate it every time.

Updated PrintBaseClass to use only one System.out.println statement.

Working on ServerServer to generate a list of actions from the database.