IoTActions are now saved/retrieved to/from the database. Now using log4j.
ServerServer now generates the action list dynamically.
For JavaIoTActions, ServerServer will first check the db for the action, if its there then it'll use the UUID.
If its not there, it will generate a UUID and put it there.

For ScriptIoTActions, ServerServer will look for each file in the script directory and try to find an action in the db.
If it exists, it'll make sure its still valid, and then add it to the official list.
If it doesn't exist, it'll create a new action and add it to an alternative list which houses actions waiting to be approved.

Moved logging to log4j.

Fixed PrintBaseClass to actually print the new line.

IoTActionDAO can not get ScriptIoTAction by name.

IoTActionList can now get the index of actions by uuid.
IoTActionList can now remove actions by uuid.